Mengerjakan fitur 7, 10, dan 11 (#30)
menyelasaikan fitur 7,10,11

+++ sifactory/src/main/java/apap/TugasAkhir/siFactory/repository/RequestUpdateItemDb.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import apap.TugasAkhir.siFactory.model.RequestUpdateItemModel;
 
 @Repository
-public interface RequestUpdateItemDb extends JpaRepository<RequestUpdateItemModel, Long> {
+public interface RequestUpdateItemDb extends JpaRepository<RequestUpdateItemModel, Integer> {
 }
